move webengine variants to their own file
